requirements.txt
fastapi
uvicorn
gunicorn

python-dotenv
python-multipart

openai
langchain
langchain-openai
pydantic
tenacity
tiktoken

numpy
scikit-learn
nltk

python-docx
pypdf
PyPDF2

rich
loguru
typer

redis

pytest
pytest-mock
